requirements.txt
# Dev environment
pip
autopep8

# App
google-generativeai # Client library for the Gemini API
gradio # Web UI library
Pillow # PIL image-loading library